# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/items": {
      "post": {
        "description": "Persists the item, emits an `item_submitted` event, and enqueues it on the `items` queue.",
        "operationId": "ItemController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitItemDto"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SubmitItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created item.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessTimestampDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ItemDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Submit an item for processing",
        "tags": [
          "items"
        ]
      },
      "get": {
        "operationId": "ItemController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Rows per page.",
            "schema": {
              "minimum": 1,
              "maximum": 600,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ItemSortField"
            }
          },
          {
            "name": "sort_direction",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of non-deleted items. `metadata.pagination.total` is the row count of the returned page.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessTimestampDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ItemListEntryDto"
                          }
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/PaginationMetadataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List items with their processing status",
        "tags": [
          "items"
        ]
      }
    }
  },
  "info": {
    "title": "H3 Zoom Test API",
    "description": "Item submission and asynchronous processing API.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "SuccessTimestampDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "ContentType": {
        "type": "string",
        "enum": [
          "text",
          "long_text",
          "numeric",
          "file"
        ]
      },
      "ItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              }
            ]
          },
          "label": {
            "type": "string"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "nullable": true
          },
          "file_ref": {
            "type": "string",
            "nullable": true
          },
          "mime_type": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "content_type",
          "label",
          "value",
          "file_ref",
          "mime_type",
          "size",
          "created_at",
          "updated_at",
          "deleted_at"
        ]
      },
      "SubmitItemDto": {
        "type": "object",
        "properties": {
          "content_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              }
            ]
          },
          "label": {
            "type": "string",
            "example": "Quarterly report"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Required for text/long_text (string) and numeric (number); omitted for file."
          },
          "file_ref": {
            "type": "string"
          },
          "mime_type": {
            "type": "string",
            "example": "application/pdf"
          },
          "size": {
            "type": "number",
            "example": 20480
          }
        },
        "required": [
          "content_type",
          "label"
        ]
      },
      "ItemStatus": {
        "type": "string",
        "enum": [
          "pending",
          "done"
        ]
      },
      "ItemResultDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "example": 87
          }
        },
        "required": [
          "score"
        ]
      },
      "ItemListEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              }
            ]
          },
          "label": {
            "type": "string"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "nullable": true
          },
          "file_ref": {
            "type": "string",
            "nullable": true
          },
          "mime_type": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemStatus"
              }
            ]
          },
          "result": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemResultDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "content_type",
          "label",
          "value",
          "file_ref",
          "mime_type",
          "size",
          "created_at",
          "status",
          "result"
        ]
      },
      "PaginationInfoDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "size": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "page",
          "size",
          "total"
        ]
      },
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
      "SortInfoDto": {
        "type": "object",
        "properties": {
          "by": {
            "type": "string",
            "example": "created_at"
          },
          "direction": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SortDirection"
              }
            ]
          }
        },
        "required": [
          "by",
          "direction"
        ]
      },
      "PaginationMetadataDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfoDto"
          },
          "sort": {
            "$ref": "#/components/schemas/SortInfoDto"
          }
        },
        "required": [
          "pagination",
          "sort"
        ]
      },
      "ItemSortField": {
        "type": "string",
        "enum": [
          "created_at",
          "updated_at",
          "label"
        ]
      }
    }
  }
}